Move redeem fragment to fragment package
